+++ libs/cardano-ledger-core/src/Cardano/Ledger/Language.hs
@@ -48,6 +48,7 @@ import NoThunks.Class (NoThunks)
 data Language
   = PlutusV1
   | PlutusV2
+  | PlutusV3
   deriving (Eq, Generic, Show, Ord, Enum, Bounded, Ix)
 
 instance NoThunks Language
@@ -78,10 +79,12 @@ instance FromJSONKey Language where
 languageToText :: Language -> Text
 languageToText PlutusV1 = "PlutusV1"
 languageToText PlutusV2 = "PlutusV2"
+languageToText PlutusV3 = "PlutusV3"
 
 languageFromText :: MonadFail m => Text -> m Language
 languageFromText "PlutusV1" = pure PlutusV1
 languageFromText "PlutusV2" = pure PlutusV2
+languageFromText "PlutusV3" = pure PlutusV3
 languageFromText lang = fail $ "Error decoding Language: " ++ show lang
 
 instance EncCBOR Language where
@@ -99,6 +102,7 @@ instance ToExpr Language
 data SLanguage (l :: Language) where
   SPlutusV1 :: SLanguage 'PlutusV1
   SPlutusV2 :: SLanguage 'PlutusV2
+  SPlutusV3 :: SLanguage 'PlutusV3
 
 deriving instance Eq (SLanguage l)
 
@@ -115,6 +119,7 @@ fromSLanguage :: SLanguage l -> Language
 fromSLanguage = \case
   SPlutusV1 -> PlutusV1
   SPlutusV2 -> PlutusV2
+  SPlutusV3 -> PlutusV3
 
 -- | For implicit reflection on '@SLanguage@'
 -- See "Cardano.Ledger.Alonzo.TxInfo" for example usage
@@ -126,6 +131,9 @@ instance IsLanguage 'PlutusV1 where
 
 instance IsLanguage 'PlutusV2 where
   isLanguage = SPlutusV2
+
+instance IsLanguage 'PlutusV3 where
+  isLanguage = SPlutusV3
 
 toSLanguage :: forall l m. (IsLanguage l, MonadFail m) => Language -> m (SLanguage l)
 toSLanguage lang