+++ eras/alonzo/impl/CHANGELOG.md
@@ -14,6 +14,8 @@
 * Made it possible for `FromJSON` to decode `CostModels` both as the new approach:
   1. as a list of cost models values,
   2. and the old approach of mapping from the parameter name to the cost model value
+* Added support for Plutus V3 in the types and functions that use `Language`.
+  (Note that the Alonzo ledger era rules do not allow V3 scripts, however.).
 
 ###`testlib`